Выносит функцию в отдельный модуль
